lint: Add the code inside tools/ to the checks.
The scripts inside tools/ weren't being checked by the linter because they
don't have the ".js" extension.

They have been added to the files passed to ESLint by Gulp, and all the
violations have been fixed, with the exception of two undefined elements
in tools/mockPayload (which needs updating) and the obfuscated data in
tools/setup (which has been excluded in linter).