refactoring: Refactor getReferences utility method.
Previously, the findKeywords method was used to find phrases
containing keywords that would reference a certain issue. However,
this method failed to take into account whether the referenced number
referred to an open issue instead of a closed issue or a PR. This
resulted in false "no references in commits" alarms for several PRs.

To resolve this, we merge the findKeywords method into the getReferences
method and refactor it to check for issue status.

Fixes #155.