templates: Add list of steps to new contributor message.
The goal here is to encourage folks not to claim an issue until they
are ready.

We're counting on the contributor not to accept the GitHub invite until
ready, which is not ideal because (a) they might just accept it, and
(b) there's a risk that the invite will expire. So we'd probably be better
off splitting this into two interactions (joining repo; claiming issue).

This also shortens the message if the initial invite hasn't been accepted.